pyproject.toml
[project]
name = "hormonal-bot"
version = "0.1.0"
description = "Телеграм бот для отслеживания менструального цикла с уведомлениями"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.23.0",
    "sqlalchemy>=2.0.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "greenlet>=3.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

